Improves contact info and cart item display
Enhances the display of contact information in the footer by fetching fallback values from environment variables if the API returns empty data.

Updates the cart sidebar and product details to provide a clearer, more user-friendly presentation of cart items including item details, size and color and total price of items. This includes changing the size selector to a dropdown.

Fixes a bug where the company info was not being displayed in the website if there were any API errors.