Improves login flow and error handling
Refactors the login process to handle different server responses (201 and 202) more effectively.

Instead of throwing an error for a 202 status (which indicates SMS validation is required), the user is redirected to the SMS validation page.

Adds a reset to the form after submission. Improves error handling by displaying toast messages for errors returned from the API.